ipfailover keepalived split brain
OSE 3.3 introduced an error in how the keepalived config file is
created. This problem persisted in OSE 3.4 as well. This change restores
the OSE 3.2 operation to OSE 3.3 and 3.4. OSE 3.5 has a different
implemenation that does not have this problem.

When multiple VIPs are configured the OSE 3.2 version places the first
in a vrrp_sync_group and ignores the other VIPs. OSE 3.3 and 3.4 place
all of the VIPs in the same vrrp-sync-group. Neither approach is
correct.  However, since vrrp-sync-group is not needed and is
eliminated in OSE 3.5 the OSE 3.2 approach does work. Placing all of
the VIPs in a single vrrp-group is the source of the problem in OSE
3.3 and OSE 3.4.

The workaround is to use the OSE 3.2 ipfailover image in OSE 3.3 and 3.4
installations. This change eliminates the need for the workaround. The
referenced bug contains the discussion.

bug 1402536
https://bugzilla.redhat.com/show_bug.cgi?id=1402536

Signed-off-by: Phil Cameron <pcameron@redhat.com>